face_detector/requirements.txt
jupyter
notebook
matplotlib
opencv-python
pillow
requests
AWSIoTPythonSDK